+++ src/main/java/edu/icet/service/admin/AdminService.java
@@ -7,10 +7,10 @@ import java.util.List;
 
 public interface AdminService {
      boolean addAdmin (Admin admin);
-     boolean deleteAdmin (Integer adminID);
-     boolean updateAdmin(Integer adminID,Admin admin);
-     Admin getAdminById(Integer adminID);
-     boolean adminExists(Integer adminID);
+     boolean deleteAdmin (Long adminID);
+     boolean updateAdmin(Long adminID,Admin admin);
+     Admin getAdminById(Long adminID);
+     boolean adminExists(Long adminID);
      Integer countAdmins();
      List<Admin>getAllAdmins();
      List<Admin>getAdminByType(AdminType type);
